Section4-Outils en Entreprise/requirements.txt
ollama==0.6.2
pillow==12.2.0